Refactor: activity -> fragment로 변경
