+++ app/src/main/java/com/diary/paintlog/view/fragments/DiaryFragment.kt
@@ -1,0 +1,24 @@
+package com.diary.paintlog.view.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.diary.paintlog.databinding.FragmentDiaryBinding
+
+class DiaryFragment : Fragment() {
+
+    private var _binding: FragmentDiaryBinding? = null // 바인딩 객체 선언
+    private val binding get() = _binding!! // 바인딩 객체 접근용 getter
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentDiaryBinding.inflate(inflater, container, false) // 바인딩 객체 초기화
+
+
+        return binding.root
+    }
+}
