+++ app/src/main/res/layout/fragment_main.xml
@@ -1,0 +1,95 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <com.prolificinteractive.materialcalendarview.MaterialCalendarView
+        android:id="@+id/calendar_view"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingHorizontal="10dp"
+        android:theme="@style/CalenderViewCustom"
+        app:layout_constraintTop_toTopOf="parent"
+        app:mcv_dateTextAppearance="@style/CalenderViewDateCustomText"
+        app:mcv_firstDayOfWeek="monday"
+        app:mcv_showOtherDates="all"
+        app:mcv_weekDayTextAppearance="@style/CalenderViewWeekCustomText" />
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/topic_view"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginTop="15dp"
+        app:layout_constraintTop_toBottomOf="@id/calendar_view">
+
+        <TextView
+            android:id="@+id/topic"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:fontFamily="@font/bm_yeon_sung"
+            android:text="추천 주제"
+            android:textColor="@color/deep"
+            android:textSize="20dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <ImageButton
+            android:id="@+id/topic_canvas"
+            android:layout_width="match_parent"
+            android:layout_height="180dp"
+            android:background="@drawable/canvas_small_selector"
+            app:layout_constraintStart_toStartOf="@id/topic"
+            app:layout_constraintTop_toBottomOf="@+id/topic"
+            />
+
+        <TextView
+            android:id="@+id/today_topic"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:fontFamily="@font/bm_yeon_sung"
+            android:text="내가 가장 좋아하는 음악은?"
+            android:textColor="@color/deep"
+            android:textSize="20dp"
+            app:layout_constraintLeft_toLeftOf="@id/topic_canvas"
+            app:layout_constraintRight_toRightOf="@id/topic_canvas"
+            app:layout_constraintTop_toTopOf="@id/topic_canvas"
+            app:layout_constraintBottom_toBottomOf="@id/topic_canvas"/>
+
+        <ImageButton
+            android:id="@+id/repeat_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="15dp"
+            android:layout_marginEnd="15dp"
+            android:background="@drawable/repeat_selector"
+            app:layout_constraintEnd_toEndOf="@id/topic_canvas"
+            app:layout_constraintTop_toTopOf="@+id/topic_canvas" />
+
+        <ImageButton
+            android:id="@+id/add_diary_button"
+            android:layout_width="70dp"
+            android:layout_height="70dp"
+            android:layout_marginTop="15dp"
+            android:background="@drawable/add_diary_selector"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/topic_canvas" />
+
+        <TextView
+            android:id="@+id/addDiaryText"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:fontFamily="@font/bm_yeon_sung"
+            android:text="새 일기"
+            android:textColor="@color/main"
+            android:textSize="20dp"
+            app:layout_constraintTop_toBottomOf="@id/add_diary_button"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
